requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
pydantic>=2.11.0
resend>=2.0.0
gradio>=5.49.0
fastapi>=0.120.0
tavily-python>=0.5.0

# Stock Research Agent dependencies
yfinance>=0.2.40
httpx>=0.27.0

# Phase 2: UI/UX enhancements
fpdf2>=2.7.0
